Fix bare exception and a buggy check
I ran into a problem with this check, I think it's occurring because the dataset is installed via pip and not `setup.py develop`. It was failing because  `self.__class__` was returning `<class 'lexibank_allenbai.Dataset'>`, instead of `Dataset` (or precisely <class 'pylexibank.dataset.Dataset'>`) so there was a mro glitch. 

Using `isinstance` should fix this and I've also added a better error message rather than a bare exception. 

BUT: I'm not sure