Support "lexibank dataset from existing CLDF" scenario
To support integrating datasets with other lexibank datasets when
they have not been created using pylexibank, we need to drop some
assumptions on how the CLDF data will look like.

This mostly means merging data based on propertyUrls of the respective
columns, not just column names. But we also accept a metadata file
name of `Wordlist-metadata.json` now, rather than the default
`cldf-metadata.json`.

For backwards compatibility, but also for readability, we stick with
default CLDF column names for the db - rather than using the propertyUrl.

closes #66